pyproject.toml
[project]
name = "contracting-smtp-automation"
version = "1.0"
requires-python = "==3.12.*"
dependencies = [
    "pandas==2.2.3",
    "google-cloud-bigquery==3.19.0",
    "google-cloud-logging==3.13.0",
    "google-cloud-storage==3.8.0",
    "click==8.3.1",
    "colorlog==6.10.1",
    "jinja2==3.1.6",
    "weasyprint==67.0",
    "resend==2.21.0",
]

[tool.ruff]
line-length = 120
indent-width = 4

# Assume Python 3.7
target-version = "py37"
exclude = ["app/scheduled_tasks/pick_back_up.py"]


[tool.ruff.lint]
# The way the Flask app is setup currently means
# that there are a ton of unused imports, we want
# to ignore these holistically for now
extend-ignore = ["F401", "E203", "E266", "E402", "E501", "E711", "E712", "F403"]
select = ["E", "F", "I"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
